Update ikea.rst
It is possible to pair and use multiple control devices for one group of lights. Next control devices has to be paired not directly with gateway, but by pressing pairing buttons on first and next control devices. Instructions in Tradfri application are incorrect, but this method works and both control devices appears on list of devices and works correctly.